Working on service unit testing.

+++ server/src/main/java/org/craftercms/profile/v2/services/impl/AuthenticationServiceImpl.java
@@ -16,13 +16,13 @@
  */
 package org.craftercms.profile.v2.services.impl;
 
-import org.craftercms.commons.crypto.SimpleDigest;
+import org.craftercms.commons.crypto.CipherUtils;
 import org.craftercms.commons.mongo.MongoDataException;
-import org.craftercms.commons.security.permissions.annotations.HasPermission;
+import org.craftercms.commons.security.exception.ActionDeniedException;
+import org.craftercms.commons.security.permissions.PermissionEvaluator;
 import org.craftercms.commons.security.permissions.annotations.SecuredObject;
 import org.craftercms.profile.api.Profile;
 import org.craftercms.profile.api.TenantActions;
-import org.craftercms.profile.api.TenantPermission;
 import org.craftercms.profile.api.Ticket;
 import org.craftercms.profile.api.exceptions.ProfileException;
 import org.craftercms.profile.api.services.AuthenticationService;
@@ -30,6 +30,7 @@ import org.craftercms.profile.api.services.ProfileService;
 import org.craftercms.profile.v2.exceptions.BadCredentialsException;
 import org.craftercms.profile.v2.exceptions.DisabledProfileException;
 import org.craftercms.profile.v2.exceptions.I10nProfileException;
+import org.craftercms.profile.v2.permissions.Application;
 import org.craftercms.profile.v2.repositories.TicketRepository;
 import org.springframework.beans.factory.annotation.Required;
 
@@ -41,16 +42,21 @@ import java.util.Date;
  *
  * @author avasquez
  */
-@HasPermission(type = TenantPermission.class, action = TenantActions.AUTHENTICATE)
 public class AuthenticationServiceImpl implements AuthenticationService {
 
     public static final String ERROR_KEY_AUTHENTICATE_ERROR =       "profile.auth.authenticateError";
     public static final String ERROR_KEY_GET_TICKET_ERROR =         "profile.auth.getTicketError";
     public static final String ERROR_KEY_INVALIDATE_TICKET_ERROR =  "profile.auth.invalidateTicket";
 
+    protected PermissionEvaluator<Application, String> permissionEvaluator;
     protected TicketRepository ticketRepository;
     protected ProfileService profileService;
     protected int ticketMaxAge;
+
+    @Required
+    public void setPermissionEvaluator(PermissionEvaluator<Application, String> permissionEvaluator) {
+        this.permissionEvaluator = permissionEvaluator;
+    }
 
     @Required
     public void setTicketRepository(TicketRepository ticketRepository) {
@@ -80,13 +86,14 @@ public class AuthenticationServiceImpl implements AuthenticationService {
             if (!profile.isEnabled()) {
                 throw new DisabledProfileException(profile.getId().toString(), tenantName);
             }
-            if (!comparePasswords(profile.getPassword(), password)) {
+            if (!CipherUtils.matchPassword(profile.getPassword(), password)) {
                 // Invalid password
                 throw new BadCredentialsException();
             }
 
             Ticket ticket = new Ticket();
-            ticket.setUserId(profile.getId().toString());
+            ticket.setTenant(tenantName);
+            ticket.setProfileId(profile.getId().toString());
             ticket.setTimestamp(new Date());
 
             ticketRepository.save(ticket);
@@ -98,10 +105,12 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     }
 
     @Override
-    public Ticket getTicket(@SecuredObject String tenantName, String ticketId) throws ProfileException {
+    public Ticket getTicket(String ticketId) throws ProfileException {
         try {
             Ticket ticket = ticketRepository.findById(ticketId);
             if (ticket != null) {
+                checkManageTicketsPermission(ticket.getTenant());
+
                 Calendar expirationTime = Calendar.getInstance();
                 expirationTime.setTime(ticket.getTimestamp());
                 expirationTime.add(Calendar.SECOND, ticketMaxAge);
@@ -116,23 +125,23 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     }
 
     @Override
-    public void invalidateTicket(@SecuredObject String tenantName, String ticketId) throws ProfileException {
+    public void invalidateTicket(String ticketId) throws ProfileException {
         try {
-            ticketRepository.removeById(ticketId);
+            Ticket ticket = ticketRepository.findById(ticketId);
+            if (ticket != null) {
+                checkManageTicketsPermission(ticket.getTenant());
+
+                ticketRepository.removeById(ticketId);
+            }
         } catch (MongoDataException e) {
             throw new I10nProfileException(ERROR_KEY_INVALIDATE_TICKET_ERROR, ticketId);
         }
     }
 
-    protected boolean comparePasswords(String storedPswd, String providedPswd) {
-        int idxOfSep = storedPswd.indexOf(ProfileServiceImpl.PASSWORD_SEP);
-        String storedHash = storedPswd.substring(0, idxOfSep);
-        String salt = storedPswd.substring(idxOfSep + 1);
-        SimpleDigest digest = new SimpleDigest();
-
-        digest.setBase64Salt(salt);
-
-        return storedHash.equals(digest.digestBase64(providedPswd));
+    protected void checkManageTicketsPermission(String tenantName) {
+        if (!permissionEvaluator.isAllowed(tenantName, TenantActions.MANAGE_PROFILES)) {
+            throw new ActionDeniedException(TenantActions.MANAGE_PROFILES, tenantName);
+        }
     }
 
 }
